refactor: 重新设计 TimePicker 组件
