+++ src/BootstrapBlazor/Components/TimePicker/TimePicker.razor
@@ -2,5 +2,20 @@
 @inherits ValidateBase<TimeSpan>
 
 <div @attributes="@AdditionalAttributes" class="@ClassString">
-    <TimePickerBody Value="@Value" OnConfirm="OnInternalValueChanged"></TimePickerBody>
+    <div class="time-picker-body">
+        <TimePickerCell @bind-Value="@CurrentTime" ViewMode="TimePickerCellViewMode.Hour" />
+        <TimePickerCell @bind-Value="@CurrentTime" ViewMode="TimePickerCellViewMode.Minute" />
+        @if (HasSeconds)
+        {
+            <TimePickerCell @bind-Value="@CurrentTime" ViewMode="TimePickerCellViewMode.Second" />
+        }
+    </div>
+    <div class="time-picker-footer">
+        <button type="button" class="time-panel-btn cancel" @onclick="@OnClickClose">@CancelButtonText</button>
+        <button type="button" class="time-panel-btn confirm" @onclick="@OnClickConfirm">@ConfirmButtonText</button>
+    </div>
 </div>
+
+@code {
+
+}
