+++ src/BootstrapBlazor/Components/TimePicker/TimePicker.razor.scss
@@ -1,9 +1,101 @@
 .bb-time-picker {
-    color: #606266;
-    border: 1px solid #ced4da;
-    box-shadow: 0 2px 12px 0 rgba(0,0,0,.1);
-    background: #fff;
+    --bb-time-picker-width: 180px;
+    background-color: var(--bs-body-bg);
+    border: var(--bs-border-width) solid var(--bs-border-color);
+    box-shadow: 0 2px 12px 0 rgba(0, 0, 0, .1);
     border-radius: var(--bs-border-radius);
-    display: inline-block;
-    padding: 1rem;
+    width: var(--bb-time-picker-width);
+    user-select: none;
+
+    .time-picker-body {
+        display: flex;
+        flex-wrap: nowrap;
+        justify-content: space-around;
+        width: 100%;
+    }
+
+    .time-picker-footer {
+        border: var(--bs-border-width) solid var(--bs-border-color);
+        padding: 4px;
+        height: 36px;
+        line-height: 25px;
+        text-align: right;
+
+        .time-panel-btn {
+            border: none;
+            line-height: 28px;
+            padding: 0 5px;
+            margin: 0 5px;
+            cursor: pointer;
+            background-color: transparent;
+            outline: none;
+            font-size: 12px;
+        }
+    }
+}
+
+.time-panel {
+    background-color: var(--bs-body-bg);
+    border: var(--bs-border-width) solid var(--bs-border-color);
+    box-shadow: 0 2px 12px 0 rgba(0,0,0,.1);
+    border-radius: var(--bs-border-radius);
+    width: 180px;
+    user-select: none;
+
+    .time-panel-content {
+        font-size: 0;
+        position: relative;
+        overflow: hidden;
+
+        &:before {
+            padding-left: 50%;
+            margin-right: 12%;
+            margin-left: 12%;
+        }
+
+        &:after {
+            left: 50%;
+            margin-left: 12%;
+            margin-right: 12%;
+        }
+
+        &.has-seconds {
+            &:before {
+                padding-left: 33.33333%;
+            }
+
+            &:after {
+                left: 66.66667%;
+                pointer-events: none;
+            }
+        }
+
+        &.havenot-seconds {
+            &:before {
+                padding-left: 50%;
+            }
+
+            &:after {
+                left: 50%;
+                pointer-events: none;
+            }
+        }
+
+        &:after, &:before {
+            content: "";
+            top: 50%;
+            position: absolute;
+            margin-top: -36.6px;
+            height: 36.6px;
+            left: 0;
+            right: 0;
+            border-top: var(--bs-border-width) solid var(--bs-border-color);
+            border-bottom: var(--bs-border-width) solid var(--bs-border-color);
+        }
+    }
+}
+
+.time-spinner-input.input .input-inner, .time-spinner-list {
+    padding: 0;
+    text-align: center;
 }
