+++ src/BootstrapBlazor/Components/TimePicker/TimePickerCell.razor
@@ -5,7 +5,7 @@
 <div @attributes="@AdditionalAttributes" class="@ClassString" id="@Id">
     <i class="@UpIconString" @onclick="OnClickUp"></i>
     <i class="@DownIconString" @onclick="OnClickDown"></i>
-    <div class="time-spinner-list-wrapper">
+    <div class="time-spinner-body">
         <ul class="time-spinner-list" style="@StyleName">
             <li class="time-spinner-item">&nbsp;</li>
             @foreach (var index in Range)
