+++ src/BootstrapBlazor/Components/Timer/Timer.razor
@@ -22,6 +22,6 @@
     }
     else
     {
-        <TimePickerBody Value="@Value" OnConfirm="@OnStart" ConfirmButtonText="@StarText" />
+        <TimePicker Value="@Value" OnConfirm="@OnStart" ConfirmButtonText="@StarText" />
     }
 </div>
