+++ src/BootstrapBlazor/wwwroot/scss/components.scss
@@ -94,7 +94,7 @@
 @import "../../Components/Timeline/Timeline.razor.scss";
 @import "../../Components/Timer/Timer.razor.scss";
 @import "../../Components/TimePicker/TimePicker.razor.scss";
-@import "../../Components/TimePicker/TimePickerBody.razor.scss";
+@import "../../Components/TimePicker/TimePickerCell.razor.scss";
 @import "../../Components/Toast/ToastContainer.razor.scss";
 @import "../../Components/Toggle/Toggle.razor.scss";
 @import "../../Components/Tooltip/Tooltip.razor.scss";
